refactor product api into service
